requirements.txt
Django==5.0.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-cors-headers==4.3.1

# GeoDjango dependencies - sistema ya tiene GDAL instalado
# GDAL==3.6.2
